% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_secret.R
\name{add_secret}
\alias{add_secret}
\title{add_secret:}
\usage{
add_secret(
  name,
  value = rstudioapi::askForPassword(),
  users = "sambold",
  vault = "~/.vault"
)
}
\arguments{
\item{name}{String, Secret-Name}

\item{value}{String Secret-Value}

\item{users}{String, User fuer den/die Secret zugaenglich}

\item{vault}{String, Pfad zum Vaul}
}
\description{
Wrapper fuer secret::add_secret Speichert ein Secret im Vault
}
