% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_user.R
\name{add_user}
\alias{add_user}
\title{add_user:}
\usage{
add_user(user = "sambold", pub_key = "~/.ssh/id_rsa.pub", vault = "~/.vault")
}
\arguments{
\item{user}{String, User-Name}

\item{pub_key}{String, Pfad zum Public Key}

\item{vault}{String, Pfad zum Vaul}
}
\description{
Wrapper fuer secret::add_user. Legt einen User im Vault an
}
