% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_age}
\alias{calc_age}
\title{calc_age: berechnet Alter}
\usage{
calc_age(von, bis = Sys.Date(), unit = "age")
}
\arguments{
\item{von}{String, der in POSIXlt umgewandelt werden kann bzw. Date}

\item{bis}{String, der in POSIXlt umgewandelt werden kann bzw. Date,
Standardwert = Sys.Date();}

\item{unit}{String, Standardwert = age; gibt an, in welcher Einheit die
Differenz angegeben werden soll. 
Zur Auswahl stehen secs, mins, hours, weeks, months, age}
}
\value{
Alter
}
\description{
berechnet Alter bzw. verstrichene Zeit in Sekunden, Minuten, Stunden, Wochen 
oder in Monaten zwischen zwei Datumswerten.
}
\examples{
calc_age(von="2004-05-12",bis=Sys.time())
}
