% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_image.R
\name{check_for_image}
\alias{check_for_image}
\title{check_for_images:}
\usage{
check_for_image(
  opts = NULL,
  image = "selenium/standalone-firefox-debug:latest"
)
}
\arguments{
\item{opts}{List, Standardwert NULL; wenn NULL wird opts ueber die Funktion
set_auto_sudo() befuellt. Dabei werden zwei Listelemente zugewiesen
\itemize{
    \item opt: gibt an, ob Passwort automatisch ("-Ks") oder manuell ("") 
        im Terminal eingegeben werden muss
    \item sudo_pswd: Passwort zur Bestaetigung, wenn opt="-Ks"
}}

\item{image}{String, Pfad zum Image, das geprueft werden soll}
}
\value{
Boolean
}
\description{
prueft ueber Terminal-Input, ob ein Image schon vorhanden ist.
}
