% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ftp.R
\name{check_ftp}
\alias{check_ftp}
\title{check_ftp: Daten von FTP-Server abrufen}
\usage{
check_ftp(
  server = "",
  fpath.ftp = "",
  out.dir = paste0(getwd(), "/"),
  typ = "dir",
  select = F,
  cmp.dir = getwd(),
  cmp.filter = ".*\\\\.csv",
  regex = "",
  lbound = 86400,
  blacklist = NA,
  retry = F,
  sleep = 3600,
  max.retry = 24,
  verbose = F
)
}
\arguments{
\item{server}{String, erforderlich, FTP-Server (z.B. 193.58.211.75)}

\item{fpath.ftp}{String, zu durchsuchender Ordner auf dem FTP-Server ohne
Begrenzung durch "/" (z.B. daapfl)}

\item{out.dir}{String, Ausgabepfad für Downloads (Standard: paste0(getwd(),"/"))}

\item{typ}{String, Angabe wie die Auswahl der Download-Files erfolgen soll: 
über ein Referenzverzeichnis (dir), anhand eines Zeitpunktes (intime) oder 
durch einen Regex-Filter (regex). Mögliche Ausprägungen: dir, intime, regex}

\item{select}{Boolean, gibt an, ob Download-Files aus der Liste potentieller 
Downlaods (eingeschränkt durch die Auswahl über typ) zusätzlich manuell 
selektiert werden können sollen (Standard: FALSE)}

\item{cmp.dir}{String, nur relevant, wenn  typ=dir, Angabe eines 
Referenzverzeichnisses das zur Auswhal der Download-Files dient}

\item{cmp.filter}{String, nur relevant, wenn  typ=dir,
Regex-Filter mit dem Liste der Files im Referenzverzeichnis
eingeschränkt werden kann (Standard: ".*\\.csv" -> der Regex-Filter wird so
angewandt, dass immer nur Dateibezeichnungen aber nicht evtl. Pfade zu 
Unterordnern berücksichtigt werden. Dh. der oben angeführte Filter wird in 
der Funktion so ergänzt, dass 'home/data/test.csv' auf 'test.csv' reduziert 
wird. Beim Abgleich mit dem FTP-Server wird dann die Datei 'test.csv' vom 
Download ausgeschlossen. Unabhängig davon in welchem Ordner sie sich dort
befindet.)}

\item{regex}{String, nur relevant, wenn typ=regex, Regex-Filter mit dem die 
zu downloadenden Files am FTP-Server eingeschränkt werden können}

\item{lbound}{Integer, nur relevant, wenn typ=intime, Anzahl der Sekunden, die 
höchstens vergangen sein dürfen, seit ein File am FTP-Server erstellt wurde.
Dh. für einen Download muss das Erstelldatum der Datei am FTP-Server 
> (Sys.time()-lbound) sein. (Standard: 86400; Erstelldatum max. vor 24h)}

\item{blacklist}{String, Regex-Filter, Dateinamen am FTP-Server, die diesem 
Muster entsprechen, werden vom Download ausgeschlossen (Standard: NA -> kein blacklist-Filter)}

\item{retry}{Boolean, gibt an, ob automatisch mehrere Verbindungen nacheinander
zum FTP-Server aufgebaut werden sollen, bis eine zu downloadende Datei gefunden
wird bzw. bis die Anzahl maximaler Wiederholungen erreicht wurde (Standard: FALSE)}

\item{sleep}{Integer, nur relevant, wenn retry=TRUE, gibt an wie viele Sekunden
zwischen den Verbindungsversuchen gewartet werden soll (Standard: 3600)}

\item{max.retry}{Integer, nur relevant, wenn retry=TRUE, gibt an wie oft 
ein erneuter Verbindungsversuch zum FTP-Server durchgeführt werden soll
(Standard: 24)}

\item{verbose}{Boolean, gibt an, ob ausführlichere Informationen zum 
Verbindungsaufbau ausgegeben werden sollen (Standard: FALSE)}
}
\value{
kein Rückgabewert
}
\description{
Ruft ausgewählte Dateien von einem FTP-Server ab und speichert sie in einem
Zielordner. Die Auswahl der zu downloadenden Dateien kann dabei 
\itemize {
  \item anhand eines Abgleichs mit einem Referenzverzeichnis erfolgen. Hier 
    werden nur jene Dateien abgezogen, die nicht schon im Referenzverzeichnis
    vorhanden sind (der Abgleich erfolgt über den Dateinamen)
  \item anhand des Zeitpunkts definiert werden. Hier werden alle Dateien 
    abgezogen, die nach einem bestimmten Zeitpunkt am FTP-Server erstellt wurden
  \item anhand eines Regex-Filters definiert werden. Hier werden alle Dateien
    abgezogen, deren Dateiname dem festgelegten Regex-Muster entsprechen.
 }
 Aus der Liste zu downloadender Dateien, die nach Angabe einer der oben 
 beschriebenen Filteroptionen identifiziert wurde, können optional auch manuell
 jene Dateien ausgewählt werden, die abgezogen werden sollen. Ebenso können 
 Dateien durch einen Regex-Filter vom Download ausgeschlossen werden.
 
 Sofern bei einem FTP-Aufruf keine neuen Dateien gefunden wurden, besteht die
 Möglichkeit den Aufruf in bestimmten Abständen für eine bestimmte Anzahl zu 
 wiederholen.
}
