% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_join.R
\name{coalesce_join}
\alias{coalesce_join}
\title{coalesce_join: join, der Inahlte gemeinsamer Variablen miteinander vereint}
\usage{
coalesce_join(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  join = dplyr::left_join,
  keep = F,
  ...
)
}
\arguments{
\item{x}{data.frame (siehe dplyr)}

\item{y}{data.frame (siehe dplyr)}

\item{by}{character/named list, Variablen ueber die gejoint wird (siehe dplyr)}

\item{suffix}{charakter, Suffix fuer Variablennamen, die in beiden Datensaetzen
vorkommen (siehe dplyr)}

\item{join}{Funktionsaufruf fuer den Join nach dem Schema
dplyr::left_join, dplyr::full_join, etc.}

\item{keep}{Boolean, Standard FALSE; gibt an, ob die urspruenglichen Variablen,
die in beiden Datensaetzen vorkommen behalten werden sollen oder ob nur
die vereinten Variablen behalten werden sollen}

\item{...}{weitere, zu uebergebende Argumente (siehe dplyr)}
}
\value{
gejointen Datensatz
}
\description{
Beim Join koennen verschiedene Join-Varianten des dplyr-packages verwendet
werden. Variabeln, die in beiden zu joinenden Quellen vorkommen, werden 
miteinander verschmolzen. Dabei wird grundsaetzlich der Wert aus x verwendet,
außer dieser ist missing, dann wird das missing mit dem Wert aus y verwendet.
}
