% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_quos_text.R
\name{conv_quos_text}
\alias{conv_quos_text}
\title{conv_quos_text: Umwandlung von Quotations}
\usage{
conv_quos_text(qs)
}
\arguments{
\item{qs}{mit ... uebergebene quos bzw. rlang::quos}
}
\value{

}
\description{
wandelt unterschiedliche Arten von Quotations in einen String um. Umgewandelt
werden sowohl quos, ellipsis, normale quo und quo, denen ein Listelement
uebergeben wird.
}
\examples{
conv_quos_text(rlang::quos(x,y,z))
conv_quos_text(rlang::quo(x))
conv_quos_text(rlang::quo(list(x,y,z)))
# Achtung: Operatoren werden nicht in String uebernommen
conv_quos_text(rlang::quo(x <- 1))
}
