% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_size.R
\name{conv_size}
\alias{conv_size}
\title{conv_size: konvertiert Fontsize von Theme in size-Einheit}
\usage{
conv_size(x)
}
\arguments{
\item{x}{Fontsize (analog to Theme), die in size-Einheit konvertiert werden soll}
}
\value{
konvertierte Fontsize
}
\description{
oder so aehnlich ... was genau wie und wo ist mir noch nicht ganz klar. Aber
damit sollte die Schriftgroesse definiert durch theme und die Schriftgroesse
definiert ueber z.b. geom_text gleich/vergleichbar sein
}
\examples{
conv_size(14)
}
