% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_path.R
\name{crt_path}
\alias{crt_path}
\title{crt_path: Verzeichnis anlegen}
\usage{
crt_path(x)
}
\arguments{
\item{x}{Character, anzulegender Pfad}
}
\value{
Boolean, gibt an, ob Verzeichnis erfolgreich angelegt wurde oder nicht.
}
\description{
versucht Verzeichnis anzulegen und gibt Statusinfo zurueck.
}
