% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_vault.R
\name{crt_vault}
\alias{crt_vault}
\title{crt_vault}
\usage{
crt_vault(vault_path = "~/.vault")
}
\arguments{
\item{vault_path}{String, Pfad zum Vault}
}
\description{
Wrapper fuer secret::create_vault. Legt einen Vault an
}
