% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euro_format.R
\name{euro_format}
\alias{euro_format}
\title{euro_format: wandelt uebergebene Zahlenwerte ins Euro-Format um}
\usage{
euro_format(x, digits = 0, g_check = T)
}
\arguments{
\item{x}{Numeric, Zahlenwert, der in Euro-Zahlenformat umgewandelt werden soll}

\item{digits}{Numeric, Standard = 0; Anzahl der Nachkommastellen}

\item{g_check}{Boolean, Standardwert = TRUE; uebergebene NAs werden in 'n.a.'
umformatiert. Wenn FALSE, werden NAs durch '' angegeben.}
}
\value{
Character, in Euro-Zahlenformat umgewandelter Inputwert
}
\description{
Zahlenwerte werden ins Euro-Format umgewandelt. Dabei werden Tausender-Trennzeichen
und Komma-Angaben nach europaeischem Standard verwendet. Zudem wird ein €-Symbol
der Zahl vorangestellt. Wenn ein Character-Wert uebergeben wird, wird lediglich
ein €-Symbol vorangestellt. Ansonsten wird der Character-String nicht weiter
formatiert.
}
