% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mail.R
\name{fetch_mail}
\alias{fetch_mail}
\title{fetch_mail: ruft eMail ab}
\usage{
fetch_mail(
  host = "imap.gmail.com",
  port = 993,
  user = "samboldsky@gmail.com",
  pswd = bagR::get_secret("google"),
  search.from = NULL,
  search.subj = NULL,
  search.body = NULL,
  wait = 1,
  max.msg = NULL,
  java.path = "/usr/lib/jvm/java-8-openjdk-amd64/"
)
}
\arguments{
\item{host}{Character, Standard = imap.gmail.com; Angabe zum verwendeten Host}

\item{port}{Numeric, Standard = 993; Angabe zum verwendeten Port}

\item{user}{Character, Standard = samboldsky@gmail.com; User-eMail}

\item{pswd}{Character, Standard bagR::get_secret("google"); User-Passwort}

\item{search.from}{Character, Standard NULL; wenn NULL, werden alle Absender
abgerufen - ansonsten auf bestimmte Absender einschraenkbar}

\item{search.subj}{Character, Standard NULL; wenn NULL, werden alle Betreffe
abgerufen - ansonsten auf bestimmten Betreff einschraenkbar}

\item{search.body}{Character, Standard NULL; wenn NULL, werden alle Inhalte
abgerufen - ansonsten auf bestimmte Inhalte einschraenkbar}

\item{wait}{Numeric, Standard = 1, aktuell nicht verwendet (vermutlich wars 
fuer mehrfaches Abrufen gedacht bzw. als Pause dazwischen)}

\item{max.msg}{Numeric, Standard = NULL, wenn NULL werden alle Nachrichten
abgerufen. Ansonsten einschraenkbar auf die n aktuellsten Nachrichten.}

\item{java.path}{Character, Pfad zu Java (falls nicht automatisch erkannt wird)}
}
\value{
...
}
\description{
ruft eMail ueber rJython ab
}
