% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max_grpval.R
\name{find_max_grpval}
\alias{find_max_grpval}
\title{find_max_grpval: gibt je Gruppe den Wert mit dem haufigsten Vorkommen zurueck}
\usage{
find_max_grpval(x, .key, .val)
}
\arguments{
\item{x}{Inputdaten}

\item{.key}{Variable, Gruppe innerhalb der ein Wert gefunden werden soll}

\item{.val}{Variable, Auspraegungen fuer die ein Max-Wert gesucht wird}
}
\value{
tibble, in dem jedem .key genau ein Wert (mit dem haufigsten Auftreten)
    zugeordnet ist
}
\description{
Bei gleich haufigem Vorkommen wird der Wert mit der niedrigsten Sortierreihenfolge
genommen.
}
