% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_param}
\alias{get_param}
\title{get_param: erstellt aus den uebergebenen cronR-Argumenten eine Parametertabelle
#'}
\usage{
get_param(filter_cols = c("FREQ_CRONJOB", "FNAME_CRONJOB"), delim = "@@")
}
\arguments{
\item{filter_cols}{String, Standard c("FREQ_CRONJOB","FNAME_CRONJOB");
Variablennamen als String, die nicht in der Parametertabelle vorkommen
(aber als col_label vorhanden sind). Mit dem Wert NULL werden keine Filter
gesetzt}

\item{delim}{String, Standard @; Delimiter durch den die Spaltennamen 
getrennt sidn}
}
\value{
Parametertabelle
}
\description{
get_param: erstellt aus den uebergebenen cronR-Argumenten eine Parametertabelle
#'
}
