% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_running_container.R
\name{get_running_container}
\alias{get_running_container}
\title{get_running_container:}
\usage{
get_running_container(opts = NULL)
}
\arguments{
\item{opts}{List, Standardwert NULL; wenn NULL wird opts ueber die Funktion
set_auto_sudo() befuellt. Dabei werden zwei Listelemente zugewiesen
\itemize{
    \item opt: gibt an, ob Passwort automatisch ("-Ks") oder manuell ("") 
        im Terminal eingegeben werden muss
    \item sudo_pswd: Passwort zur Bestaetigung, wenn opt="-Ks"
}}
}
\value{
String
}
\description{
listet die Namen laufender Container auf
}
