% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secret.R
\name{get_secret}
\alias{get_secret}
\title{get_secret:}
\usage{
get_secret(name, key = "~/.ssh/id_rsa", vault = "~/.vault")
}
\arguments{
\item{name}{String, Secret-Name}

\item{key}{String, Pfad zum RSA-Key}

\item{vault}{String, Pfad zum Vault}
}
\value{
secret
}
\description{
Wrapper fuer secret::get_secret. Holt ein Secret aus dem Vault
}
