% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_file_enc.R
\name{guess_file_enc}
\alias{guess_file_enc}
\title{guess_file_enc}
\usage{
guess_file_enc(
  fpath,
  what = "character",
  sep = ";",
  quote = "\\"",
  skipNul = T,
  nline = 10,
  quiet = T,
  ...
)
}
\arguments{
\item{fpath}{String, Pfad (inkl. Dateiname)}

\item{what}{String, Typ dessen, was eingelesen werden soll (siehe scan)}

\item{sep}{String, Trennzeichen in Daten (siehe scan)}

\item{quote}{String, Set an Quoting-Symbolen (siehe scan)}

\item{skipNul}{Boolean, Nuls auslassen, wenn von Character eingelesen wird (siehe scan)}

\item{nline}{Integer, maximale Anzahl an Zeilen, die eingelesen werden (siehe scan)}

\item{quiet}{Boolean, Info zum Einleseprozess (siehe scan)}

\item{...}{zusaetzliche Parameter für die Funktion scan}
}
\value{
data.frame mit Liste möglicher Encodings und deren Wahrscheinlichkeit
}
\description{
Durch Angabe eines Pfades (inkl. Dateiname) wird ein Teil eines File als 
Rohdaten mit 'scan' eingelesen, um zu versuchen, das verwendete Encoding 
zu bestimmen.
}
