% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_packages.R
\name{init_packages}
\alias{init_packages}
\title{init_packages: installiert Packages}
\usage{
init_packages(pkg_list = NULL, force_install = FALSE)
}
\arguments{
\item{force_install}{Boolean, Standardwert = FALSE, gibt an, ob Installation
auch dann stattfinden soll, wenn Package bereits vorhanden}

\item{pkge_list}{String, Standardwert = NULL, Liste zu installierender Packages
wenn NULL, werden Standardpackages installiert}
}
\value{
Vektor mit Installationsstatus der einzelnen Packages: TRUE = Installation
    hat funktioniert, FALSE = Installation fehlgeschlagen
}
\description{
Installiert eine Liste von Packages. Werden keine Packages angegeben, wird
eine Liste von Standardpackages installiert. Dabei kann angegeben werden, ob
nur Packages installiert werden sollen, die nicht schon vorhanden sind oder
ob eine (Neu)installation erzwungen werden soll.
}
