% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_selenium.R
\name{init_selenium}
\alias{init_selenium}
\title{init_selenium:}
\usage{
init_selenium(
  image = "selenium/standalone-firefox-debug:latest",
  profile_path = "/home/samuel/.mozilla/firefox/testR",
  selenium_port = 4445,
  vinagre_port = NULL,
  opts = NULL
)
}
\arguments{
\item{image}{String, Pfad zum Image mit dem der Browser gestartet werden soll}

\item{profile_path}{String, Pfad zu einem bestehenden Browserprofil}

\item{selenium_port}{Numeric, Standardwert 4445, Port der fuer Selenium 
verwendet werden soll.}

\item{vinagre_port}{Numeric, Standardwert NULL, Port der fuer Vinagre 
verwendet werden soll. Wenn NULL wird vinagre nicht verwendet.}

\item{opts}{List, Standardwert NULL; wenn NULL wird opts ueber die Funktion
set_auto_sudo() befuellt. Dabei werden zwei Listelemente zugewiesen
\itemize{
    \item opt: gibt an, ob Passwort automatisch ("-Ks") oder manuell ("") 
        im Terminal eingegeben werden muss
    \item sudo_pswd: Passwort zur Bestaetigung, wenn opt="-Ks"
}}
}
\value{
RSelenium remoteDriver
}
\description{
Startet neuen Selenium Browser und schafft Vorbedingungen.
}
