% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill_container.R
\name{kill_container}
\alias{kill_container}
\title{kill_container:}
\usage{
kill_container(opts = NULL, container = NULL)
}
\arguments{
\item{opts}{List, Standardwert NULL; wenn NULL wird opts ueber die Funktion
set_auto_sudo() befuellt. Dabei werden zwei Listelemente zugewiesen
\itemize{
    \item opt: gibt an, ob Passwort automatisch ("-Ks") oder manuell ("") 
        im Terminal eingegeben werden muss
    \item sudo_pswd: Passwort zur Bestaetigung, wenn opt="-Ks"
}}

\item{container}{String, Name eines Containers, der beendet werden soll. 
Alternativ zu einem String, kann auch eine Liste mit Strings uebergeben
werden. Wenn NULL uebergeben wird, werden alle laufenden Container beendet.}
}
\value{
Information zu beendeten Containern
}
\description{
beendet bestimmte(n) Container
}
