% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_container.R
\name{list_container}
\alias{list_container}
\title{list_container:}
\usage{
list_container(opts = NULL)
}
\arguments{
\item{opts}{List, Standardwert NULL; wenn NULL wird opts ueber die Funktion
set_auto_sudo() befuellt. Dabei werden zwei Listelemente zugewiesen
\itemize{
    \item opt: gibt an, ob Passwort automatisch ("-Ks") oder manuell ("") 
        im Terminal eingegeben werden muss
    \item sudo_pswd: Passwort zur Bestaetigung, wenn opt="-Ks"
}}
}
\value{
String
}
\description{
Listet alle laufenden Container auf
}
