% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_format.R
\name{num_format}
\alias{num_format}
\title{num_format: europaeisches Zahlenformat}
\usage{
num_format(x, digits = 0, g_check = T)
}
\arguments{
\item{x}{Numeric, Wert, der in europaeisches Zahlenformat umformatiert werden soll}

\item{digits}{Numeric, Standardwert = 0, Anzahl der anzuzeigenden Nachkommastellen}

\item{g_check}{Boolean, Standardwert = TRUE, gibt an, ob uebergebene NAs als
'n.a.' (TRUE) oder als '' (FALSE) dargestellt werden sollen}
}
\value{
String, in europaeisches Format umgeformter Input
}
\description{
wandelt Zahl in europaeisches Zahlenformat um: Tausendertrennzeichen wird als
Punkt, Komma als Beistrich dargestellt. Zudem wird ein Euro-Symbol vorangestellt.
Wird ein numerischer Wert uebergeben, kommt es zu diesen Umformatierungen. Wird
ein String uebergeben, wird lediglich ein Euro-Symbol vorangestellt. Der 
String-Wert bleibt dabei unformatiert.
}
