% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_up.R
\name{round_up}
\alias{round_up}
\title{round_up: rundet Zahl zur naechsten Obergrenze hin auf}
\usage{
round_up(x, step = 10)
}
\arguments{
\item{x}{Numeric, aufzurundender Zahlenwert/-vektor}

\item{step}{Numeric, Standard 10; Schrittweite innerhalb der aufgerundet werde soll}
}
\value{
aufgerundete Zahl bzw. aufgerundeter Vektor
}
\description{
rundet eine Zahl oder einen Vektor zur naechsten Obergrenze hin auf. Als 
Obergrenzenschritt kann eine beliebige Ganzzahl definiert werden. Wird eine
Schrittlaenge von 10 angegeben, wird zum naechsten Zehner aufgerundet. Wird
eine Schrittlaenge von 5 angegeben, wird zum naechsten Fuenfer aufgerundet etc.
}
\examples{
round_up(1)
round_up(c(1,5,12))
round_up(3,step=5)
round_up(105,step=1000)
}
