% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail.R
\name{send_mail}
\alias{send_mail}
\title{send_mail}
\usage{
send_mail(
  from = "samboldsky@gmail.com",
  to = NULL,
  body = "",
  subject = "",
  files = NULL,
  user = NULL,
  pswd = secret::get_secret("google"),
  host = "smtp.gmail.com",
  port = 587
)
}
\arguments{
\item{from}{String, Sender\*in-eMail}

\item{to}{String-Vektor, eine oder mehrere Empfänger\*innen-Adressen. Wenn
NULL, wird die from-Adresse übernommen.}

\item{body}{String, Text des eMails}

\item{subject}{String, Betreff des eMails}

\item{files}{String-Vektor, Pfad für die zu versendenden Anhänge}

\item{user}{String, Username für den SMTP-Login. Wenn NULL, wird die 
from-Adresse übernommen.}

\item{pswd}{String, Passwort für den SMTP-Login}

\item{host}{String, Hostadresse für die SMTP-Verbindung}

\item{port}{Integer, Port für die SMTP-Verbindung}

\item{java.path}{String, Pfad zu der zu verwendenden Java-Version. Wird nur 
verwendet, wenn JAVA_HOME noch kein Pfad zugewiesen wurde.}
}
\value{
versendet ein eMail
}
\description{
Versendet ein eMail mit/ohne Anhang an eine\*einen oder mehrere Empfänger\*innen.
}
