% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_auto_sudo.R
\name{set_auto_sudo}
\alias{set_auto_sudo}
\title{set_auto_sudo:}
\usage{
set_auto_sudo(auto_sudo = TRUE, sudo_pswd = NULL)
}
\arguments{
\item{auto_sudo}{Boolean, Standardwert TRUE; gibt an, ob Sudo-Eingabe
automatisch erfolgen soll ("-kS") oder manuell ("").}

\item{sudo_pswd}{String, wenn Sudo-Eingabe automatisch erfolgen soll, wird
nach einem Passwort gefragt, das dann hinterlegt wird.}
}
\value{
Liste mit Sudo-Zusatz fuer Terminal und Passwort
}
\description{
legt fest, ob Sudo-Eingabe automatisch erfolgen soll und 
    hinterlegt ggf. ein Passwort
}
