% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vault.R
\name{set_vault}
\alias{set_vault}
\title{set_vault}
\usage{
set_vault(path = "~/.vault")
}
\arguments{
\item{path}{String, Pfad zum Vault}
}
\value{
Pfad zum Vault (nur bei Ausgabe in Variable)
}
\description{
Wrapper für secret::create_vault, um vault-Pfad im Environment abzulegen
}
