% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_vinagre.R
\name{show_vinagre}
\alias{show_vinagre}
\title{show_vinagre:}
\usage{
show_vinagre(vinagre_port = 5901)
}
\arguments{
\item{vinagre_port}{Numeric, gibt Port an auf dem Vinagre-Instanz laeuft}
}
\description{
oeffnet Vinagre
}
